+++ test/test_interface.py
@@ -167,8 +167,8 @@ def test_config_file_codeactions_options():
 
 
 def test_version_update_pypi():
-    from fortls.langserver import LangServer
     from fortls.jsonrpc import JSONRPC2Connection, ReadWriter
+    from fortls.langserver import LangServer
 
     parser = commandline_args("fortls")
     args = parser.parse_args("-c f90_config.json".split())